+++ Project01/src/com/project/first/Trail.java
@@ -18,7 +18,7 @@ public class Trail extends GameObject{
 	
 	// life = 0.001 - 0.1
 	
-	public Trail(int x, int y, ID id, Color color, int width, int height, float life, Handler handler) {
+	public Trail(float x, float y, ID id, Color color, int width, int height, float life, Handler handler) {
 		super(x, y, id);
 		this.handler = handler;
 		this.color = color;
